✅ 🎨 Improve file structure and finish passing test
