+++ src/setup-db.ts
@@ -1,0 +1,9 @@
+import { PrismaClient } from '@prisma/client'
+
+let prisma: PrismaClient
+
+export function startDb(): PrismaClient {
+  prisma = new PrismaClient()
+  prisma.$connect
+  return prisma
+}
